chore(ui): extracts reduceFieldMap from where builder
